Adds CNI and Network Constructs Support (#3)
* Adds CNI and Network Constructs Support

Similar to BESS, this patch virtualizes a network pipeline as a series
of modules (Network Functions) connected together via gates. This
general concept will be used to define Network Pipelines in the Nimbess
Agent, regardless of what Data Plane is used underneath. It is the
driver's responsibility to translate pipeline related requests into Data
Plane specific actions.

This patch supports a per-port pipeline construct which is generic to
any driver and managed by Nimbess Agent. Initially the support added is
for basic Switching support with CNI requests. Later the pipeline will
be stored in ETCD, but for now is maintained in Agent runtime.