Adds FastPath network support to CNI (#12)
* Adds FastPath network support to CNI

When a CNI request is made to add a kernel port for a Pod, a second
FastPath port will created as well in a secondary default FastPath
network. This port driver in the future may support multiple high speed
datapath technologies, but for now only supports DPDK. Pods placed in
the same network on their kernel port (original CNI request) will also
be placed in the same FastPath network.

Following the same methodology as userspace CNI, when the FastPath CNI
is invoked, it will annotate the Pod with configuration data and path
information so that the Pod will have access to DPDK port information as
well path to vhost sockets via the downward API. The vhost socket
directory is for now shared amongst all containers and in the future
will be created dynamically by an admission controller. See
k8s/pod_example.yaml for an example of how to deploy a Pod.

Other Changes Include:
 - Fix for an issue where ports were not being deleted after a CNI Del
   operation.
 - Added pod name/ns to protobuf for a future nimbess-CNI to include in
   CNI Request. This avoids having to search for pod name for
   annotation purposes.
 - Changed from govendor to use go modules. This is more consistent with
   Kubernetes and resolves compatibility issues with importing k8s code.
 - nimbess DS is now ran as a the nimbess service account. A new yaml
   file is provided for creating the service account and proper RBAC
   permissions. This will be re-used for stargazer.

Signed-off-by: Tim Rozet <trozet@redhat.com>