Adds etcd to Agent
Now the Agent will have access to the etcd data store setup by
stargazer. This will be used to access data written by stargazer, as
well as the data store used by the Agent. This patch adds an
initialization function which will write a simple Agent model into the
database to detect if the Agent is starting for the first time or not.

Future patches will implement writing networking related data models
into the database.

Additionally, configuration is added for etcd to the Agent config. The
etcd information will be populated by config map in k8s with the
nimbess.yaml update. Minor fix of adding Kubeconfig also included.

Signed-off-by: Tim Rozet <trozet@redhat.com>